deploy/roles/common/files/content/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==5.0a1
dbus-python==1.2.8
idna==2.7
netaddr==0.7.19
pbkdf2==1.3
PyRIC==0.1.6.3
python-networkmanager==2.1
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
termcolor==1.1.0
urllib3==1.23
wifi==0.3.8
